+++ src/base/b-slider/test/helpers.js
@@ -13,7 +13,7 @@
  */
 
 const
-	h = include('tests/helpers');
+	h = include('tests/helpers').default;
 
 /**
  * Emulates the `once` swipe on a slider