+++ src/super/i-static-page/modules/ss-helpers/page.js
@@ -7,7 +7,7 @@
  */
 
 const
-	config = require('config'),
+	config = require('@config/config'),
 	{src, webpack, csp} = config;
 
 const